Audio Optimized
BACKGROUD audio clip removed and BACKGROUND_01 load type changed to streaming and compression format to Vorbis. THis combination uses the least amount of memory. Sound effects changed to Decompress On Load and uses PCM compression Format as no decompression is needed and audio clip is short so it will load very quickly.